\name{France54Map}
\alias{France54Map}
\docType{data}
\title{
Map of the departments of France in 1954
}
\description{
A polygon map of the French departments
}
\usage{data(France54Map)}
\format{
  An sf object with 90 observations on the following 2 variables
  \describe{
    \item{\code{Dept}}{Department name}
    \item{\code{geometry}}{list of department polygons}
  }
}

\details{
This shape file is used in Chapter 7, "Re-viewing Bertin's main example", and combined with the data in the file F1954.  Combining the six new departments of 1967 into the two former departments of Seine and Seine-et-Oise is approximately right.
}

\source{
\url{ http://coulmont.com/cartes/rcarto.pdf}  Derived from GEOFLADept_FR_Corse_AV_L93/DEPARTEMENT.SHP
}

\keyword{datasets}
