\name{GmooG}
\alias{GmooG-package}
\docType{package}
\title{
GmooG: datasets analysed in "Getting (more out of) Graphics"
}

\description{
There are 25 chapters of graphical data analyses in the book.  Datasets that are not readily available are mainly provided in this package.
}

\details{
Other datasets are analysed in the book as well.  They are available in various R packages.  Some can be downloaded and updated from the web.
}

\author{
Antony Unwin unwin@math.uni-augsburg.de

}


